WIP: UTIL_EST: sched/{core,fair}: add support to use estimated utilization
The CPU/Task estimated utilization can be used for in all the code paths
interested into a signal which is more stable than PELT.

This patch refactor the __cpu_util()/cpu_util() methods to return the MAX
between the CPU estimated utilization and the current value of util_avg.
This new signal is usable both when tasks are RUNNABLE as well as when
the RQ is empty. In this last case, the blocked utilization accounted for
by the PELT signal, can still be used to estimated the "potential" load
of a currently empty CPU.
The task_util() method is refactored as well to return the task estimated
utilization.

The new behaviour is available only when the UTIL_EST feature is enabled
and for all the call-site which do not explicitly require the usage of
the PELT signal. The signal to be used is defined using an additional
parameter for the refactored methods:
  UTIL_AVG: force usage of PELT signal, i.e. util_avg
  UTIL_EST: use estimated utilization when possible,
            i.e. UTIL_EST feature enabled

By default, this patch updates all the call-sites of the refactored methods
to force the usage of the PELT signal thus not changing the current behaviour
of the CFS scheduler.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>